+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
             android:name=".feature.product.detail.ProductDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".feature.main.MainActivity"
+            android:name=".feature.product.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
