+++ app/src/main/java/woowacourse/shopping/common/ViewType.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.list
+package woowacourse.shopping.common
 
 enum class ViewType {
     PRODUCT,
