+++ app/src/main/java/woowacourse/shopping/feature/cart/CartActivity.kt
@@ -9,7 +9,6 @@ import woowacourse.shopping.R
 import woowacourse.shopping.data.cart.CartDao
 import woowacourse.shopping.data.cart.CartRepositoryImpl
 import woowacourse.shopping.databinding.ActivityCartBinding
-import woowacourse.shopping.list.adapter.CartProductListAdapter
 import woowacourse.shopping.model.CartProductState
 
 class CartActivity : AppCompatActivity(), CartContract.View {
