+++ app/src/main/java/woowacourse/shopping/feature/product/recent/RecentProductItemViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.list.viewholder
+package woowacourse.shopping.feature.product.recent
 
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewbinding.ViewBinding
