+++ app/src/main/java/woowacourse/shopping/feature/product/recent/RecentProductWrapperViewHolder.kt
@@ -1,9 +1,8 @@
-package woowacourse.shopping.list.viewholder
+package woowacourse.shopping.feature.product.recent
 
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewbinding.ViewBinding
 import woowacourse.shopping.databinding.ItemRecentProductListBinding
-import woowacourse.shopping.list.adapter.RecentProductListAdapter
 
 class RecentProductWrapperViewHolder(
     private val binding: ViewBinding,
